// ClientApi/src/app/shared/models/user.model.ts
// export class User {
//     userId : Number;
//     fullName: string;
//     email: string;
//     password: string;
//     retypePassword: string;
//     panNumber: string;
//     bank: string;
//     bankAccountNumber: Number;
// }

export class User {
    Id: string;
    FullName: string;
    Email: string;
    Password: string;
    ConfirmPassword: string;
    PanNumber: string;
    Bank: string;
    BankAccountNumber: string;
}